pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goh"
version = "0.1.0"
description = "Guilds of Heroes — D&D social network"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0,<4",
    "gunicorn>=21.2,<23",
    "click>=8.1,<9",
    "pydantic>=2.5,<3",
    "pydantic-settings>=2.1,<3",
    "structlog>=24.1,<25",
    "PyJWT>=2.8,<3",
    "bcrypt>=4.1,<5",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-cov>=4.1,<6",
    "httpx>=0.27,<1",
    "ruff>=0.3,<1",
    "mypy>=1.8,<2",
]

[project.scripts]
goh = "cli.main:cli"

[tool.setuptools.packages.find]
include = ["goh*", "cli*", "api*", "config*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["goh", "cli", "api", "config"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
